Update dependencies of strapi-plugin-email-designer
The versions of several dependencies in the package.json file of strapi-plugin-email-designer have been updated, in order to keep up with the latest improvements and bug fixes of these packages. This includes FontAwesome, Day.js, html-to-text, React email editor, Babel eslint parser, Cypress linting, Prettier, and the Strapi dependency itself. This results in better performance and potentially less vulnerabilities.