Change templating language to Mustache in README.md
The README.md for the 'strapi-plugin-email-designer' plugin has been updated to reflect a switch from lodash to mustache for email templating. Mustache was chosen for its efficient and cleaner syntax. This commit updates the instructions and examples in the readme to correctly show users how to use mustache templating in their email design, increasing readability and easy user adoption.